+++ include/gameLogick/StartScreen.hpp
@@ -18,8 +18,8 @@ class StartScreen : public Screen {
     sf::Color btnBG;
     sf::Color hoverBG;
     sf::Color outlineColor;
-    int outlineSize = 1;
-    int outlineHoverSize = 2;
+    float outlineSize = 1.f;
+    float outlineHoverSize = 2.f;
 
     virtual void draw(sf::RenderTarget& target, sf::RenderStates states) const;
     virtual void render(sf::Event& event);
